requirements.txt
apiclient==1.0.4
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
crayons==0.3.0
fuzzywuzzy==0.17.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
httplib2==0.17.0
idna==2.8
numpy==1.18.1
oauth2client==4.1.3
pandas==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
uritemplate==3.0.1
urllib3==1.25.8
webdriver-manager==2.3.0
